Revert to overwriting of manifest etc in Dar

+++ src/DarConverter.js
@@ -3,7 +3,7 @@ const crypto = require('crypto')
 const fs = require('fs')
 const path = require('path')
 
-const {replaceExt} = require('./helpers/util')
+const {getExt, replaceExt} = require('./helpers/util')
 const xml = require('./helpers/xml')
 
 const Converter = require('./Converter')
@@ -40,14 +40,12 @@ class DarConverter extends Converter {
       let converter
       if (type === 'sheet') {
         converter = new SheetMLConverter()
-      } else if (type === 'article') {
-        converter = new JATSConverter()
       } else {
-        throw new Error(`Unhandled document type "${type}" when importing Dar`)
+        converter = new JATSConverter()
       }
-      const doc = await converter.import(path.join(darPath, filePath), volume, options)
+      const exedoc = await converter.import(path.join(darPath, filePath), volume, options)
 
-      files[filePath] = doc
+      files[filePath] = exedoc
     }
 
     return {
@@ -56,12 +54,12 @@ class DarConverter extends Converter {
     }
   }
 
-  async export (doc, darPath, volume = fs, options = {}) {
+  async export (exedoc, darPath, volume = fs, options = {}) {
     let files
-    if (doc.type === 'Folder') {
-      files = doc.files
+    if (exedoc.type === 'Folder') {
+      files = exedoc.files
     } else {
-      files = {'unnamed': doc}
+      files = {'unnamed': exedoc}
     }
 
     const manifest = xml.load(`
@@ -74,10 +72,21 @@ class DarConverter extends Converter {
     let documentsEl = manifest('documents')
 
     for (let [fileName, node] of Object.entries(files)) {
+      const ext = getExt(fileName)
+
       let type
-      let filePath
+      let filePath = fileName
       let converter
-      if (node.body && node.body.length === 1 && node.body[0].type === 'Table') {
+
+      // TODO this is a temporary implementation to avoid overwriting files
+      // already 'managed' in the manifest. More consideration needs to
+      // be given to how this is best done (i.e overwriting manifest or 'merging' other files into it;
+      // a general option to turn on/off overwriting of already converted files?)
+      if (ext === 'jats.xml') {
+        type = 'article'
+      } else if (ext === 'sheet.xml') {
+        type = 'sheet'
+      } else if (node.body && node.body.length === 1 && node.body[0].type === 'Table') {
         type = 'sheet'
         filePath = replaceExt(fileName, 'sheet.xml')
         converter = new SheetMLConverter()
@@ -85,12 +94,10 @@ class DarConverter extends Converter {
         type = 'article'
         filePath = replaceExt(fileName, 'jats.xml')
         converter = new JATSConverter()
-      } else {
-        throw new Error(`Unhandled node type "${node.type}" when exporting to Dar`)
       }
 
-      if (converter) {
-        let filePathFull = path.join(darPath, filePath)
+      let filePathFull = path.join(darPath, filePath)
+      if (converter && !volume.existsSync(filePathFull)) {
         await converter.export(node, filePathFull, volume)
       }
 
@@ -100,7 +107,9 @@ class DarConverter extends Converter {
     }
 
     const manifestPath = path.join(darPath, 'manifest.xml')
-    await this.writeFile(manifestPath, xml.dump(manifest), volume)
+    if (!volume.existsSync(manifestPath)) {
+      await this.writeFile(manifestPath, xml.dump(manifest), volume)
+    }
   }
 }
